feat: enhance delete functionality to check user authorization and remove associated image

+++ backend/controllers/stuff.js
@@ -1,3 +1,4 @@
+const fs = require('fs');
 // backend/controllers/stuff.js
 const Thing = require('../models/thing');
 
@@ -53,16 +54,25 @@ exports.modifyThing = (req, res, next) => {
 };
 
 exports.deleteThing = (req, res, next) => {
-  Thing.deleteOne({ _id: req.params.id })
-    .then(() => {
-      res.status(200).json({ message: 'Deleted!' });
+  Thing.findOne({ _id: req.params.id })
+    .then(thing => {
+      if (thing.userId != req.auth.userId) {
+        res.status(401).json({ message: 'Not authorized' });
+      } else {
+        const filename = thing.imageUrl.split('/images/')[1];
+        fs.unlink(`images/${filename}`, () => {
+          Thing.deleteOne({ _id: req.params.id })
+            .then(() => res.status(200).json({ message: 'Objet supprimé !' }))
+            .catch(error => res.status(401).json({ error }));
+        });
+      }
     })
-    .catch((error) => {
-      res.status(400).json({ error });
+    .catch(error => {
+      res.status(500).json({ error });
     });
 };
 
-exports.getAllStuff = (req, res, next) => {
+exports.getAllThings = (req, res, next) => {
   Thing.find()
     .then((things) => {
       res.status(200).json(things);
